+++ tests/node.rs
@@ -7,9 +7,9 @@ mod node {
     #[test]
     fn children() {
         let mut taffy = Taffy::new();
-        let child1 = taffy.new_node(Style::default(), &[]).unwrap();
-        let child2 = taffy.new_node(Style::default(), &[]).unwrap();
-        let node = taffy.new_node(Style::default(), &[child1, child2]).unwrap();
+        let child1 = taffy.new_node(FlexboxLayout::default(), &[]).unwrap();
+        let child2 = taffy.new_node(FlexboxLayout::default(), &[]).unwrap();
+        let node = taffy.new_node(FlexboxLayout::default(), &[child1, child2]).unwrap();
 
         assert_eq!(taffy.child_count(node).unwrap(), 2);
         assert_eq!(taffy.children(node).unwrap()[0], child1);
@@ -20,7 +20,7 @@ mod node {
     fn set_measure() {
         let mut taffy = Taffy::new();
         let node =
-            taffy.new_leaf(Style::default(), MeasureFunc::Raw(|_| Size { width: 200.0, height: 200.0 })).unwrap();
+            taffy.new_leaf(FlexboxLayout::default(), MeasureFunc::Raw(|_| Size { width: 200.0, height: 200.0 })).unwrap();
         taffy.compute_layout(node, Size::undefined()).unwrap();
         assert_eq!(taffy.layout(node).unwrap().size.width, 200.0);
 
@@ -32,14 +32,14 @@ mod node {
     #[test]
     fn add_child() {
         let mut taffy = Taffy::new();
-        let node = taffy.new_node(Style::default(), &[]).unwrap();
+        let node = taffy.new_node(FlexboxLayout::default(), &[]).unwrap();
         assert_eq!(taffy.child_count(node).unwrap(), 0);
 
-        let child1 = taffy.new_node(Style::default(), &[]).unwrap();
+        let child1 = taffy.new_node(FlexboxLayout::default(), &[]).unwrap();
         taffy.add_child(node, child1).unwrap();
         assert_eq!(taffy.child_count(node).unwrap(), 1);
 
-        let child2 = taffy.new_node(Style::default(), &[]).unwrap();
+        let child2 = taffy.new_node(FlexboxLayout::default(), &[]).unwrap();
         taffy.add_child(node, child2).unwrap();
         assert_eq!(taffy.child_count(node).unwrap(), 2);
     }
@@ -48,10 +48,10 @@ mod node {
     fn remove_child() {
         let mut taffy = Taffy::new();
 
-        let child1 = taffy.new_node(Style::default(), &[]).unwrap();
-        let child2 = taffy.new_node(Style::default(), &[]).unwrap();
+        let child1 = taffy.new_node(FlexboxLayout::default(), &[]).unwrap();
+        let child2 = taffy.new_node(FlexboxLayout::default(), &[]).unwrap();
 
-        let node = taffy.new_node(Style::default(), &[child1, child2]).unwrap();
+        let node = taffy.new_node(FlexboxLayout::default(), &[child1, child2]).unwrap();
         assert_eq!(taffy.child_count(node).unwrap(), 2);
 
         taffy.remove_child(node, child1).unwrap();
@@ -66,10 +66,10 @@ mod node {
     fn remove_child_at_index() {
         let mut taffy = Taffy::new();
 
-        let child1 = taffy.new_node(Style::default(), &[]).unwrap();
-        let child2 = taffy.new_node(Style::default(), &[]).unwrap();
+        let child1 = taffy.new_node(FlexboxLayout::default(), &[]).unwrap();
+        let child2 = taffy.new_node(FlexboxLayout::default(), &[]).unwrap();
 
-        let node = taffy.new_node(Style::default(), &[child1, child2]).unwrap();
+        let node = taffy.new_node(FlexboxLayout::default(), &[child1, child2]).unwrap();
         assert_eq!(taffy.child_count(node).unwrap(), 2);
 
         taffy.remove_child_at_index(node, 0).unwrap();
@@ -84,10 +84,10 @@ mod node {
     fn replace_child_at_index() {
         let mut taffy = Taffy::new();
 
-        let child1 = taffy.new_node(Style::default(), &[]).unwrap();
-        let child2 = taffy.new_node(Style::default(), &[]).unwrap();
+        let child1 = taffy.new_node(FlexboxLayout::default(), &[]).unwrap();
+        let child2 = taffy.new_node(FlexboxLayout::default(), &[]).unwrap();
 
-        let node = taffy.new_node(Style::default(), &[child1]).unwrap();
+        let node = taffy.new_node(FlexboxLayout::default(), &[child1]).unwrap();
         assert_eq!(taffy.child_count(node).unwrap(), 1);
         assert_eq!(taffy.children(node).unwrap()[0], child1);
 
@@ -100,12 +100,12 @@ mod node {
     fn remove() {
         let mut taffy = Taffy::new();
 
-        let style2 = Style { flex_direction: FlexDirection::Column, ..Style::default() };
+        let style2 = FlexboxLayout { flex_direction: FlexDirection::Column, ..FlexboxLayout::default() };
 
         // Build a linear tree layout: <0> <- <1> <- <2>
         let node2 = taffy.new_node(style2, &[]).unwrap();
-        let node1 = taffy.new_node(Style::default(), &[node2]).unwrap();
-        let node0 = taffy.new_node(Style::default(), &[node1]).unwrap();
+        let node1 = taffy.new_node(FlexboxLayout::default(), &[node2]).unwrap();
+        let node0 = taffy.new_node(FlexboxLayout::default(), &[node1]).unwrap();
 
         assert_eq!(taffy.children(node0).unwrap().as_slice(), &[node1]);
 
@@ -123,16 +123,16 @@ mod node {
     fn set_children() {
         let mut taffy = Taffy::new();
 
-        let child1 = taffy.new_node(Style::default(), &[]).unwrap();
-        let child2 = taffy.new_node(Style::default(), &[]).unwrap();
-        let node = taffy.new_node(Style::default(), &[child1, child2]).unwrap();
+        let child1 = taffy.new_node(FlexboxLayout::default(), &[]).unwrap();
+        let child2 = taffy.new_node(FlexboxLayout::default(), &[]).unwrap();
+        let node = taffy.new_node(FlexboxLayout::default(), &[child1, child2]).unwrap();
 
         assert_eq!(taffy.child_count(node).unwrap(), 2);
         assert_eq!(taffy.children(node).unwrap()[0], child1);
         assert_eq!(taffy.children(node).unwrap()[1], child2);
 
-        let child3 = taffy.new_node(Style::default(), &[]).unwrap();
-        let child4 = taffy.new_node(Style::default(), &[]).unwrap();
+        let child3 = taffy.new_node(FlexboxLayout::default(), &[]).unwrap();
+        let child4 = taffy.new_node(FlexboxLayout::default(), &[]).unwrap();
         taffy.set_children(node, &[child3, child4]).unwrap();
 
         assert_eq!(taffy.child_count(node).unwrap(), 2);
@@ -144,10 +144,10 @@ mod node {
     fn set_style() {
         let mut taffy = Taffy::new();
 
-        let node = taffy.new_node(Style::default(), &[]).unwrap();
+        let node = taffy.new_node(FlexboxLayout::default(), &[]).unwrap();
         assert_eq!(taffy.style(node).unwrap().display, Display::Flex);
 
-        taffy.set_style(node, Style { display: Display::None, ..Style::default() }).unwrap();
+        taffy.set_style(node, FlexboxLayout { display: Display::None, ..FlexboxLayout::default() }).unwrap();
         assert_eq!(taffy.style(node).unwrap().display, Display::None);
     }
 
@@ -155,9 +155,9 @@ mod node {
     fn mark_dirty() {
         let mut taffy = Taffy::new();
 
-        let child1 = taffy.new_node(Style::default(), &[]).unwrap();
-        let child2 = taffy.new_node(Style::default(), &[]).unwrap();
-        let node = taffy.new_node(Style::default(), &[child1, child2]).unwrap();
+        let child1 = taffy.new_node(FlexboxLayout::default(), &[]).unwrap();
+        let child2 = taffy.new_node(FlexboxLayout::default(), &[]).unwrap();
+        let node = taffy.new_node(FlexboxLayout::default(), &[child1, child2]).unwrap();
 
         taffy.compute_layout(node, taffy::geometry::Size::undefined()).unwrap();
 
@@ -181,8 +181,8 @@ mod node {
     fn remove_last_node() {
         let mut taffy = Taffy::new();
 
-        let parent = taffy.new_node(Style::default(), &[]).unwrap();
-        let child = taffy.new_node(Style::default(), &[]).unwrap();
+        let parent = taffy.new_node(FlexboxLayout::default(), &[]).unwrap();
+        let child = taffy.new_node(FlexboxLayout::default(), &[]).unwrap();
         taffy.add_child(parent, child).unwrap();
 
         taffy.remove(child);