+++ tests/relayout.rs
@@ -5,7 +5,7 @@ fn relayout() {
     let mut taffy = taffy::Taffy::new();
     let node1 = taffy
         .new_node(
-            taffy::style::Style {
+            taffy::style::FlexboxLayout {
                 size: taffy::geometry::Size { width: Dimension::Points(8f32), height: Dimension::Points(80f32) },
                 ..Default::default()
             },
@@ -14,7 +14,7 @@ fn relayout() {
         .unwrap();
     let node0 = taffy
         .new_node(
-            taffy::style::Style {
+            taffy::style::FlexboxLayout {
                 align_self: taffy::prelude::AlignSelf::Center,
                 size: taffy::geometry::Size { width: Dimension::Auto, height: Dimension::Auto },
                 // size: taffy::geometry::Size { width: Dimension::Percent(1.0), height: Dimension::Percent(1.0) },
@@ -25,7 +25,7 @@ fn relayout() {
         .unwrap();
     let node = taffy
         .new_node(
-            taffy::style::Style {
+            taffy::style::FlexboxLayout {
                 size: taffy::geometry::Size { width: Dimension::Percent(1f32), height: Dimension::Percent(1f32) },
                 ..Default::default()
             },